create addToCart function in cart context to replae handleClick
